Moved Debug and Version into their own namespace, componentizing them.
