+++ library/Zend/Feed/PubSubHubbub/Publisher.php
@@ -14,7 +14,7 @@ use Traversable;
 use Zend\Http\Request as HttpRequest;
 use Zend\Stdlib\ArrayUtils;
 use Zend\Uri;
-use Zend\Version;
+use Zend\Version\Version;
 
 /**
  * @category   Zend
