+++ library/Zend/Feed/PubSubHubbub/Subscriber.php
@@ -16,7 +16,7 @@ use Traversable;
 use Zend\Http\Request as HttpRequest;
 use Zend\Stdlib\ArrayUtils;
 use Zend\Uri;
-use Zend\Version;
+use Zend\Version\Version;
 
 /**
  * @category   Zend
